+++ addon/components/cardflow.js
@@ -49,7 +49,7 @@ export default class CardflowComponent extends Component {
     title: 'Catalog',
     type: 'card',
     format: 'grid',
-    component: 'cards/collection',
+    component: 'cards/master-collection',
     value: {
       id: 'catalog-card',
       type: 'catalog',
