--- addon/components/cards/collection.hbs
@@ -1,13 +1,0 @@
-{{#let @model as |item|}}
-  <Boxel @class="master-collection__boxel boxel-card">
-    <div class="master-collection">
-      <h3 class="master-collection__title">{{item.catalog_title}}</h3>
-      <div class="master-collection__desc">{{item.catalog_description}}</div>
-      <CoverArt @covers={{item.selected_art}} />
-      <div class="master-collection__count">
-        <div class="master-collection__num">{{item.number_of_songs}}</div>
-        <div>Masters</div>
-      </div>
-    </div>
-  </Boxel>
-{{/let}}
