+++ addon/components/cards/master-collection.hbs
@@ -1,0 +1,28 @@
+{{#let @model as |item|}}
+  <Boxel @class="master-collection__boxel">
+    {{#if @transition}}
+      <BoxelHighlight {{on "click" (fn @transition item)}} @class="master-collection__highlighter" @on={{item.selected}} />
+      {{#if (eq @mode "edit")}}
+        <BoxelActions
+          @displayHighlightActions={{item.selected}}
+          @btnLeft="more-actions"
+          @btnLeftIcon="more-actions"
+          @btnRightIcon="expand"
+          @btnRightAction={{fn @transition item}}
+        />
+      {{else}}
+        <IconButton {{on "click" (fn @transition item)}} @icon="expand" @class="boxel-card__expand-btn master-collection__select-btn" />
+      {{/if}}
+    {{/if}}
+    {{!-- <button {{on "click" (if @togglePin (fn @togglePin item))}} class="master-collection__pin-btn {{if item.pinned "pinned"}}" aria-label="pin">{{svg-jar "pin" width="22px" height="22px"}}</button> --}}
+    <div class="master-collection">
+      <h3 class="master-collection__title">{{item.catalog_title}}</h3>
+      <div class="master-collection__desc">{{item.catalog_description}}</div>
+      <CoverArt @covers={{item.selected_art}} />
+      <div class="master-collection__count">
+        <div class="master-collection__num">{{item.number_of_songs}}</div>
+        <div>Masters</div>
+      </div>
+    </div>
+  </Boxel>
+{{/let}}
