--- app/components/cards/collection.js
@@ -1,1 +1,0 @@
-export { default } from '@cardstack/boxel/components/cards/collection';
