+++ app/components/cards/master-collection-item.js
@@ -1,0 +1,1 @@
+export { default } from '@cardstack/boxel/components/cards/master-collection-item';
