+++ app/components/isolated-collection.js
@@ -1,0 +1,1 @@
+export { default } from '@cardstack/boxel/components/isolated-collection';
