+++ tests/dummy/app/controllers/media-registry/collection.js
@@ -1,35 +1,21 @@
 import MediaRegistryIndexController from './index';
 import { action } from '@ember/object';
 import { dasherize } from '@ember/string';
-import { isBlank } from '@ember/utils';
-
 
 export default class MediaRegistryCollectionController extends MediaRegistryIndexController {
   @action
-  selectOrTransitionToDetail(item) {
+  transitionToIsolate(item) {
     let itemId = dasherize(item.song_title.trim());
     this.transitionToRoute('media-registry.item', itemId);
   }
 
   @action
-  transitionToPrevious() {
+  transitionToView() {
     this.transitionToRoute('media-registry.collection', this.model.title);
   }
 
-  @action async search(query) {
-    let collection = this.model.collection;
-    if (isBlank(query)) {
-      return collection;
-    } else {
-      let lowerQuery = query.toLowerCase();
-      return collection.filter(i =>
-        this.model.columns.some(c =>
-           c.isSearchable !== false &&
-           c.valuePath &&
-           !isBlank(i[c.valuePath]) &&
-           i[c.valuePath].toLowerCase().includes(lowerQuery)
-        )
-      );
-    }
+  @action
+  transitionToEdit() {
+    this.transitionToRoute('media-registry.collection.edit', this.model.title);
   }
 }
