+++ tests/dummy/app/controllers/media-registry/index.js
@@ -1,40 +1,49 @@
 import Controller from '@ember/controller';
-import { action, set } from '@ember/object';
+import { action, get } from '@ember/object';
 import { dasherize } from '@ember/string';
-import { get } from '@ember/object';
-import { compare } from '@ember/utils';
+import { compare, isBlank } from '@ember/utils';
 
 export default class MediaRegistryIndexController extends Controller {
   removed = [];
 
   @action
-  togglePin(item) {
-    set(item, 'pinned', !item.pinned);
-  }
-
-  @action
-  toggleSelect(item) {
-    let collection = this.model.collection;
-    set(item, 'selected', !item.selected);
-    set(collection, 'selectedItemCount', collection.filter(item => item.selected).length);
-    set(collection, 'selectedAll', collection.length === collection.selectedItemCount);
-  }
-
-  @action
-  selectOrTransition(item) {
+  transitionToIsolate(item) {
     let itemId = dasherize(item.catalog_title.trim());
     this.transitionToRoute('media-registry.collection', itemId);
+  }
+
+  @action
+  transitionToEdit() {
+    this.transitionToRoute('media-registry.edit');
+  }
+
+  @action
+  transitionToView() {
+    this.transitionToRoute('media-registry');
+  }
+
+  @action
+  async search(query) {
+    let collection = this.model.collection;
+    if (isBlank(query)) {
+      return collection;
+    } else {
+      let lowerQuery = query.toLowerCase();
+      return collection.filter(i =>
+        this.model.columns.some(c =>
+            c.isSearchable !== false &&
+            c.valuePath &&
+            !isBlank(i[c.valuePath]) &&
+            String(i[c.valuePath]).toLowerCase().includes(lowerQuery)
+        )
+      );
+    }
   }
 
   @action
   async sort(column, direction) {
     let multiplier = (direction === 'asc') ? 1 : -1;
     return this.model.collection.sort((a, b) => multiplier * compare(get(a, column.valuePath), get(b, column.valuePath)))
-  }
-
-  @action
-  transitionToPrevious() {
-    this.transitionToRoute('media-registry');
   }
 
   @action
