+++ tests/dummy/app/styles/media-registry/collection.css
@@ -47,7 +47,6 @@
 /* Collection Item Wrapper + Actions */
 .collection__boxel {
   width: 100%;
-  height: 100%;
   border-radius: 5px;
 }
 
@@ -97,7 +96,7 @@
 /* Collection Item */
 .collection {
   display: grid;
-  grid-auto-rows: min-content;
+  grid-auto-rows: max-content;
   text-align: center;
   width: 100%;
   height: 100%;
