+++ tests/dummy/app/styles/media-registry/master-collection.css
@@ -7,7 +7,7 @@
 /* Master Collection Item Wrapper + Actions */
 .master-collection__boxel {
   width: 100%;
-  height: 100%;
+  transition: box-shadow var(--boxel-transition);
 }
 
 .master-collection__highlighter {
@@ -71,7 +71,9 @@
   grid-template-rows: auto auto 1fr;
   align-items: end;
   padding: 60px 20px 15px;
-  border-radius: 10px;
+  border-radius: var(--boxel-border-radius);
+  border: var(--boxel-border);
+  background-color: white;
 }
 
 .master-collection__title,
