+++ tests/dummy/app/templates/media-registry/collection.hbs
@@ -5,27 +5,18 @@
     <span class="breadcrumbs__pill breadcrumbs__pill--current">{{titleize (humanize @model.title)}}</span>
   </div>
 
-  <MediaCollection @model={{@model}} @sort={{this.sort}} @search={{this.search}} @transition={{this.selectOrTransitionToDetail}} @removeItem={{this.removeItem}} @mode="view" @format="table" as |item|>
-    <Boxel @class="collection__boxel">
-      <BoxelHighlight {{on "click" (fn this.selectOrTransitionToDetail item)}} @class="collection__highlighter" @on={{item.selected}} />
-      <div class="collection">
-        <div class="collection__img" style={{css-url "background-image" item.cover_art_medium}}>
-          <button class="collection__play-btn">{{svg-jar "play-btn" width="60px" height="60px"}}</button>
-        </div>
-        <h3 class="collection__title">{{titleize item.song_title}}</h3>
-        <div class="collection__artist">{{titleize item.artist}}</div>
-      </div>
-    </Boxel>
-    <div class="collection-editor__btn-group collection__select-btn {{if item.selected "collection__select-btn--selected"}}">
-      <button {{on "click" (fn this.toggleSelect item)}} class="collection-editor__select-btn" aria-label="select">
-        {{#if item.selected}}
-          {{svg-jar "icon-circle-selected" class="collection-editor__select--selected" width="16px" height="16px"}}
-        {{else}}
-          {{svg-jar "icon-circle" width="16px" height="16px"}}
-        {{/if}}
-      </button>
-    </div>
-  </MediaCollection>
+  <IsolatedCollection
+    @mode="view"
+    @model={{@model}}
+    @format="table"
+    @transition={{this.transitionToIsolate}}
+    @transitionToEdit={{this.transitionToEdit}}
+    @transitionToView={{this.transitionToView}}
+    @sort={{this.sort}}
+    @search={{this.search}}
+    @removeItem={{this.removeItem}}
+    @itemComponent="cards/master-collection-item"
+  />
 </div>
 
 {{outlet}}
