+++ tests/dummy/app/templates/media-registry/edit.hbs
@@ -1,34 +1,13 @@
-<BoxelModal @onClose={{this.transitionToPrevious}} @class="collection-edit-modal">
-  <MediaCollection @model={{@model}} @mode="edit" @format="table" @sort={{this.sort}} @transition={{this.selectOrTransition}} @removeItem={{this.removeItem}} as |item|>
-    <Boxel @class="master-collection__boxel">
-      <BoxelHighlight {{on "click" (fn this.selectOrTransition item)}} @class="master-collection__highlighter" @on={{item.selected}} />
-      <BoxelActions
-        @displayHighlightActions={{item.selected}}
-        @btnLeft="more-actions"
-        @btnLeftIcon="more-actions"
-        @btnRightIcon="expand"
-        @btnRightAction={{fn this.selectOrTransition item}}
-      />
-      {{!-- <button {{on "click" (fn this.togglePin item)}} class="master-collection__pin-btn {{if item.pinned "pinned"}}" aria-label="pin">{{svg-jar "pin" width="22px" height="22px"}}</button> --}}
-      <div class="master-collection boxel-card">
-        <h3 class="master-collection__title">{{item.catalog_title}}</h3>
-        <div class="master-collection__desc">{{item.catalog_description}}</div>
-        <CoverArt @covers={{item.selected_art}} />
-        <div class="master-collection__count">
-          <div class="master-collection__num">{{item.number_of_songs}}</div>
-          <div>Masters</div>
-        </div>
-      </div>
-    </Boxel>
-    <div class="collection-editor__btn-group">
-      <button {{on "click" (fn this.toggleSelect item)}} class="collection-editor__select-btn" aria-label="select">
-        {{#if item.selected}}
-          {{svg-jar "icon-circle-selected" class="collection-editor__select--selected" width="16px" height="16px"}}
-        {{else}}
-          {{svg-jar "icon-circle" width="16px" height="16px"}}
-        {{/if}}
-      </button>
-      <button {{on "click" (fn this.removeItem item)}} class="collection-editor__select-btn" aria-label="remove">{{svg-jar "icon-minus-circle" width="16px" height="16px"}}</button>
-    </div>
-  </MediaCollection>
+<BoxelModal @onClose={{this.transitionToView}} @class="collection-edit-modal">
+  <IsolatedCollection
+    @mode="edit"
+    @model={{@model}}
+    @transition={{this.transitionToIsolate}}
+    @transitionToEdit={{this.transitionToEdit}}
+    @transitionToView={{this.transitionToView}}
+    @sort={{this.sort}}
+    @search={{this.search}}
+    @removeItem={{this.removeItem}}
+    @itemComponent="cards/master-collection"
+  />
 </BoxelModal>
