+++ tests/dummy/public/media-registry/api/all_tracks_combined.json
@@ -1221,7 +1221,7 @@
     "cover_art_date": "2020-02-25"
   },
   {
-    "song_title": "The leaves are changing color (dance remix)",
+    "song_title": "The leaves are changing color (Dance remix)",
     "artist": "Pia Midina",
     "album": "Autumn Leaves",
     "cover_art": "media-registry/covers/Autumn-Leaves.jpg",
